requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.31
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
requests==2.32.3
langchain==0.3.0
langchain-community==0.3.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
dashscope==1.20.0